fix: circular redundancy
