+++ pyeudiw/openid4vp/vp_sd_jwt_vc.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
+from pyeudiw.jwt.helper import is_jwt_expired
 from pyeudiw.jwt.parse import KeyIdentifier_T, extract_key_identifier
-from pyeudiw.jwt.verification import is_jwt_expired
 from pyeudiw.openid4vp.exceptions import InvalidVPKeyBinding
 from pyeudiw.openid4vp.interface import VpTokenParser, VpTokenVerifier
 from pyeudiw.sd_jwt.exceptions import InvalidKeyBinding, UnsupportedSdAlg
