+++ pyeudiw/tests/jwt/test_verification.py
@@ -1,8 +1,10 @@
+from pyeudiw.jwt.helper import is_jwt_expired
 from pyeudiw.jwt.jws_helper import JWSHelper
-from pyeudiw.jwt.verification import is_jwt_expired, verify_jws_with_key
 from pyeudiw.jwk import JWK
 
 from cryptojwt.jwk.ec import new_ec_key
+
+from pyeudiw.jwt.verification import verify_jws_with_key
 
 def test_is_jwt_expired():
     jwk = new_ec_key('P-256')
